rename encoder and solver files consistently
